requirements.txt
elasticsearch==7.0.2
faker==1.0.5
mysql-connector-python==8.0.17